move viewModel to correct package (presentation)
